+++ src/utilities.cc
@@ -151,7 +151,7 @@ NAN_METHOD(format) {
   // Which load/save operations are available for each compressed format?
   Local<Object> format = New<Object>();
   for (std::string f : {
-    "jpeg", "png", "webp", "tiff", "magick", "openslide", "dz", "ppm", "fits", "gif", "svg", "pdf", "v"
+    "jpeg", "png", "webp", "tiff", "magick", "openslide", "dz", "ppm", "fits", "gif", "svg", "heif", "pdf", "v"
   }) {
     // Input
     Local<Boolean> hasInputFile =